contentblitz/requirements.txt
# ContentBlitz — multi-agent content pipeline
setuptools>=70,<82
streamlit>=1.32.0
openai>=1.40.0
faiss-cpu>=1.8.0; python_version < "3.14"
numpy>=1.22.0
python-dotenv>=1.0.0
httpx>=0.27.0
pypdf>=4.0.0
serpapi>=1.0.2
fastmcp>=3.1.1
mcp>=1.26.0
